feat(app-shell): marketplace:installed-list SDUI widget (Installed Apps body)
cloud ADR-0009 P2a: the Installed Apps page migrates to plugin-carried
metadata; the console contributes only this registered widget. The
legacy React route renders the SAME component during the migration
window — single implementation, two entries.
