+++ .changeset/installed-list-widget.md
@@ -1,0 +1,5 @@
+---
+"@object-ui/app-shell": minor
+---
+
+`marketplace:installed-list` SDUI widget — the Installed Apps body (control-plane/local dual-source list, refresh, uninstall) extracted from the React route page, which now renders the same component. The page shell ships as metadata with `@objectstack/cloud-connection`'s install-local plugin (cloud ADR-0009 P2a).
